% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc_offset_berlin.R
\name{append_utc_offset_string}
\alias{append_utc_offset_string}
\title{Append UTC Offset String to Character Timestamps}
\usage{
append_utc_offset_string(x, offsets)
}
\arguments{
\item{x}{vector of character timestamps in format yyyy-mm-dd HH:MM:SS}

\item{offsets}{vector of UTC offsets}
}
\value{
vector of character resulting from pasting UTC offset strings such
  as "+0100" for UTC+1 to the input strings in \code{x}
}
\description{
Append UTC Offset String to Character Timestamps
}
\examples{
# Timestamp in CET -> UTC+1
append_utc_offset_string("2019-01-30 10:34:00", 1)

# Timestamp in CEST -> UTC+2
append_utc_offset_string("2019-05-30 10:34:00", 2)

}
