% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{berlin_local_timestamps_to_POSIXct}
\alias{berlin_local_timestamps_to_POSIXct}
\title{Convert Berlin Local Timestamps (Text) to POSIXct}
\usage{
berlin_local_timestamps_to_POSIXct(timestamps)
}
\arguments{
\item{timestamps}{vector of character representing timestamps in format
\code{"yyyy-mm-dd HH:MM:SS"}}
}
\description{
This function creates POSIXct objects from local timestamps (character) in 
the format \code{"yyyy-mm-dd HH:MM:SS"}, of which is known that they are 
recorded in time zone "Europe/Berlin", i.e. CET in winter and CEST in summer. 
It is required that the \code{timestamps} are ordered by time, which should 
be the case if they  were recorded by a measuring device.
}
\examples{
timestamps <- c(
  "2017-10-29 01:30:00", # 1: CEST
  "2017-10-29 02:00:00", # 2: CEST
  "2017-10-29 02:30:00", # 3: CEST
  "2017-10-29 02:00:00", # 4: CET
  "2017-10-29 02:30:00", # 5: CET
  "2017-10-29 03:00:00"  # 6: CET
)

berlin_local_timestamps_to_POSIXct(timestamps)

}
