% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{berlinNormalTimeToBerlinLocalTime}
\alias{berlinNormalTimeToBerlinLocalTime}
\title{berlinNormalTimeToBerlinLocalTime}
\usage{
berlinNormalTimeToBerlinLocalTime(x)
}
\arguments{
\item{x}{character string representing a timestamp measured in Berlin without
adjusting time during the summer period, i.e. keeping the normal (= winter) 
time (= UTC+1)}
}
\description{
berlinNormalTimeToBerlinLocalTime
}
\keyword{internal}
