% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{berlinNormalTimeToUTC}
\alias{berlinNormalTimeToUTC}
\title{berlinNormalTimeToUTC}
\usage{
berlinNormalTimeToUTC(x)
}
\arguments{
\item{x}{character string representing a timestamp measured in Berlin without
adjusting time during the summer period, i.e. keeping the normal (= winter) 
time (= UTC+1)}
}
\description{
berlinNormalTimeToUTC
}
\keyword{internal}
