% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{currentDateString}
\alias{currentDateString}
\title{String representing the current Date}
\usage{
currentDateString(format = "\%d \%B \%Y")
}
\arguments{
\item{format}{format string containing percentage-placeholders as defined in 
\code{strptime}}
}
\value{
character string representing the current date
}
\description{
String representing the current Date
}
\examples{
currentDateString()
currentDateString("\%d.\%m.\%Y")
currentDateString("\%Y-\%m-\%d")

}
