% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{currentQuarter}
\alias{currentQuarter}
\title{Number of current Quarter}
\usage{
currentQuarter()
}
\value{
number of current quarter (1, 2, 3 or 4)
}
\description{
Number of current Quarter
}
\keyword{internal}
