% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeMain.R
\name{currentYear}
\alias{currentYear}
\title{Current Year (as numeric)}
\usage{
currentYear()
}
\value{
Current year as four digit number (numeric)
}
\description{
Current Year (as numeric)
}
\keyword{internal}
