% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range_cest.R
\name{date_range_CEST}
\alias{date_range_CEST}
\title{When does Summer Time start / end?}
\usage{
date_range_CEST(year)
}
\arguments{
\item{year}{Scalar year number between 1980 and 2100.}
}
\description{
At what days does the summer time start /end in a given year?
}
\examples{
# At what days does summer time start and end, respectively, in 2010?
date_range_CEST(2010)
  
# Check if summer time really starts at 2010-03-28. Timestamps between
# 2:00 (inclusive) and 3:00 (exclusive) do not exist in Central European Time
# Note that in this case R removes the time information!
as.POSIXct("2010-03-28 01:59:59", tz = "Europe/Berlin") # CET
as.POSIXct("2010-03-28 02:00:00", tz = "Europe/Berlin") # Time removed!
as.POSIXct("2010-03-28 02:59:59", tz = "Europe/Berlin") # Time removed!
as.POSIXct("2010-03-28 03:00:00", tz = "Europe/Berlin") # CEST

# Check if summer time really ends at "2010-10-31. Timestamps between
# 2:00 (inclusive) and 3:00 (exclusive) exist twice, once in CEST and a 
# second time in CET, so R does not know which one you mean! 
as.POSIXct("2010-10-31 01:00:00", tz = "Europe/Berlin") # CEST
as.POSIXct("2010-10-31 02:00:00", tz = "Europe/Berlin") # CEST

# R seems to decide (on my computer!) that times before 02:01:50 belong to
# CEST and at or after that time belong to CET!
as.POSIXct("2010-10-31 02:01:49", tz = "Europe/Berlin") # CEST
as.POSIXct("2010-10-31 02:01:50", tz = "Europe/Berlin") # CET

as.POSIXct("2010-10-31 02:02:00", tz = "Europe/Berlin") # CET
as.POSIXct("2010-10-31 03:00:00", tz = "Europe/Berlin") # CET

# Get the starts and ends of CEST for a sequence of years
date_range_CEST(2017:2020)

}
