% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeMain.R
\name{daylightSavingTimeInEffect}
\alias{daylightSavingTimeInEffect}
\title{Is Daylight Saving Time (DST) in Effect?}
\usage{
daylightSavingTimeInEffect(x, tz = "")
}
\arguments{
\item{x}{object of class POSIXt (either POSIXct or POSIXlt)}

\item{tz}{time zone string given to \code{\link{as.POSIXlt}}}
}
\value{
Returns (hopefully!) \code{TRUE} if daylight saving time is in effect
  at the given time, \code{FALSE} if daylight saving time is not in effect
  and \code{NA} if it is unknown
}
\description{
Is Daylight Saving Time (DST) in Effect?
}
\keyword{internal}
