% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{daysPerMonth}
\alias{daysPerMonth}
\title{Number of Days in the Month of the given Date}
\usage{
daysPerMonth(date)
}
\arguments{
\item{date}{date object}
}
\value{
(integer) number of days in the month of the given date
}
\description{
Number of Days in the Month of the given Date
}
\examples{
# You may either pass a Date object...
daysPerMonth(as.Date("2010-01-04"))

# ... or a date string in yyyy-mm-dd format
daysPerMonth("2010-01-04")

# Number of days in February 2010
daysPerMonth("2010-02-01")

# Number of days in February 2012
daysPerMonth("2012-02-15")

}
