% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{decade}
\alias{decade}
\title{Year Number to Decade Number}
\usage{
decade(year)
}
\arguments{
\item{year}{year number, e.g. 2016}
}
\description{
Convert a year number to a decade number (round down to full decade)
}
\keyword{internal}
