% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimestampSummary.R
\name{getTimestampSummary}
\alias{getTimestampSummary}
\title{Summary about a Sequence of (Text) Timestamps}
\usage{
getTimestampSummary(x, format = NULL, template_index = NULL)
}
\arguments{
\item{x}{vector of character representing timestamps}

\item{format}{format description, such as "\%Y-\%m-\%d \%H:\%M:\%S"}

\item{template_index}{index in \code{x} from which to select a timestamps 
that is used as a template when looking for an appropriate timestamp format}
}
\description{
Summary about a Sequence of (Text) Timestamps
}
\examples{
x <- kwb.datetime::sequenceOfTimestamps("2019-10-31", "2019-11-01")
getTimestampSummary(x)

}
