% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeStep.R
\name{getTimestepInSeconds}
\alias{getTimestepInSeconds}
\title{Get Time Step in Seconds}
\usage{
getTimestepInSeconds(timestamps, default = 60)
}
\arguments{
\item{timestamps}{vector of POSIXt objects}

\item{default}{default time step in seconds. Default: 60}
}
\value{
smallest, non-zero timestep in seconds found in the sequence of
  timestamps
}
\description{
Find the time step applied in a sequence of timestamps. Give a warning if 
more than one time step was found.
}
