% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeMain.R
\name{hsDateStr}
\alias{hsDateStr}
\title{Timestamp or Date Object to String in format yyyy-mm-dd}
\usage{
hsDateStr(tstamp)
}
\arguments{
\item{tstamp}{vector of date or time objects}
}
\description{
Timestamp or Date Object to String in format yyyy-mm-dd
}
