% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{hsDateTimeToPosix}
\alias{hsDateTimeToPosix}
\title{Date and minutes to POSIXct}
\usage{
hsDateTimeToPosix(strDate, intTime_min)
}
\arguments{
\item{strDate}{character string representing a date in format yyyy-mm-dd}

\item{intTime_min}{time since midnight of the day given in \code{strDate}}
}
\description{
Converts date given as string and time since midnight in seconds into POSIXct
object in GMT time zone.
}
\keyword{internal}
