% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{hsDateTimeToPosix2}
\alias{hsDateTimeToPosix2}
\title{Date and Seconds to POSIXct}
\usage{
hsDateTimeToPosix2(myDate, myTime_min, myExampPosix)
}
\arguments{
\item{myDate}{character string representing a date in format yyyy-mm-dd}

\item{myTime_min}{time since midnight of the day given in \code{myDate}}

\item{myExampPosix}{example POSIXct object from which the origin is 
determined}
}
\description{
Returns POSIXct object representing given date and time. In myExampPosix an
example POSIXct object must be given in order to calculate the correct 
origin.
}
\keyword{internal}
