% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeMain.R
\name{hsQua}
\alias{hsQua}
\title{Number of Month to Number to Quarter}
\usage{
hsQua(month)
}
\arguments{
\item{month}{number of month (1:12)}
}
\value{
number of quarter: 1 if month in (1:3), 2 if month in (4:6), 3 if
  month in (7:9), 4 if month in (10:12)
}
\description{
Number of Month to Number to Quarter
}
\keyword{internal}
