% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeMain.R
\name{hsQuaStr}
\alias{hsQuaStr}
\title{Timestamp to Quarter String}
\usage{
hsQuaStr(tstamp)
}
\arguments{
\item{tstamp}{(vector of) timestamp(s) of type POSXIXt}
}
\value{
string of the form \dQuote{Q<n> of <yyyy>} with <n> representing the 
  number of the quarter within the year and <yyyy> representing the year.
}
\description{
Timestamp to Quarter String
}
\keyword{internal}
