% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{hsTsAfter}
\alias{hsTsAfter}
\title{Timestamps after Timestamp Limit?}
\usage{
hsTsAfter(tstamps, tsLimit, limIncluded = TRUE)
}
\arguments{
\item{tstamps}{vector of timestamps, either as character strings in
ISO-format (yyyy-mm-dd HH:MM:SS), as Date objects or as POSIXt-objects}

\item{tsLimit}{first timestamp (\code{limIncluded} == TRUE) or timestamp
directly before first timestamp (\code{limIncluded} == FALSE) to be
considered}

\item{limIncluded}{if \code{TRUE}, timestamps that are equal to 
\code{tsLimit} are considered to be contained in the result set}
}
\description{
Returns a vector of booleans indicating whether the timestamps at 
corresponding positions in a vector of timestamps lie after a given 
timestamp.
}
\keyword{internal}
