% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{hsTsBefore}
\alias{hsTsBefore}
\title{Timestamps before Timestamp Limit?}
\usage{
hsTsBefore(tstamps, tsLimit, limIncluded = TRUE)
}
\arguments{
\item{tstamps}{vector of timestamps, either as character strings in
ISO-format (yyyy-mm-dd HH:MM:SS), as Date objects or as POSIXt-objects}

\item{tsLimit}{last timestamp (\code{limIncluded} == TRUE) or timestamp
directly after last timestamp (\code{limIncluded} == FALSE) to be
considered}

\item{limIncluded}{if \code{TRUE}, timestamps that are equal to 
\code{tsLimit} are considered to be contained in the result set}
}
\description{
Returns a vector of booleans indicating whether the timestamps at 
corresponding positions in a vector of timestamps lie before a given 
timestamp.
}
\keyword{internal}
