% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeFilter.R
\name{hsTsIn}
\alias{hsTsIn}
\title{Deprecated use \code{\link{timestampIn}} instead}
\usage{
hsTsIn(...)
}
\arguments{
\item{...}{passed to \code{\link{timestampIn}}}
}
\description{
Deprecated use \code{\link{timestampIn}} instead
}
