% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsTsInfo}
\alias{hsTsInfo}
\title{Full Information on Timestamps}
\usage{
hsTsInfo(tstamp, tzone = "", tsep = "T")
}
\arguments{
\item{tstamp}{(vector of) timestamp(s), either of POSIXt-type or character}

\item{tzone}{time zone in which timestamps in \emph{tstamp} are given. Will
only be used if \emph{tstamp} are not yet of POSIXt-type.}

\item{tsep}{character used in ISO-8601 text representation to divide date
from time}
}
\value{
data frame with columns \emph{tPosix.local} (POSIXct-objects in time 
  zone \emph{tzone}), \emph{tzone} (name of time zone), \emph{tPosix.UTC} 
  (POSIXct-objects in time zone \emph{UTC}), \emph{tChr.ISO} (timestamp in 
  ISO-8601 syntax), \emph{isdst} (1 if daylight savings time is in effect, 
  otherwise 0).
}
\description{
different representations of timestamp(s), including POSIXct-objects for 
local and UTC timezone as well as ISO-8601 compliant (text) timestamp and 
information if daylight savings time is in effect
}
\keyword{internal}
