% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsTStampsPerDay}
\alias{hsTStampsPerDay}
\title{Number of Timestamps per Day}
\usage{
hsTStampsPerDay(tstamps, tsName = "date")
}
\arguments{
\item{tstamps}{vector of POSIX-timestamps}

\item{tsName}{name of "date" column in result data frame}
}
\value{
data frame with columns \code{date} (or the name given in
  \code{tsName}) and \code{count}
}
\description{
Number of timestamps per day in a vector of POSIX-time objects
}
\keyword{internal}
