% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{localNormalTimeToLocalTime}
\alias{localNormalTimeToLocalTime}
\title{localNormalTimeToLocalTime}
\usage{
localNormalTimeToLocalTime(x, UTCOffset, tz)
}
\arguments{
\item{x}{character timestamp that is interpretable by
\code{\link{as.POSIXct}}, representing a time given in UTC}

\item{UTCOffset}{UTC offset in number of hours that the local normal time is 
ahead of UTC, e.g. +01 hour for Berlin nomral (= winter) time}

\item{tz}{time zone of the local time}
}
\description{
localNormalTimeToLocalTime
}
\keyword{internal}
