% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{localNormalTimeToUTC}
\alias{localNormalTimeToUTC}
\title{localNormalTimeToUTC}
\usage{
localNormalTimeToUTC(x, UTCOffset)
}
\arguments{
\item{x}{Character timestamp that is interpretable by
\code{\link{as.POSIXct}}, representing a time given in UTC}

\item{UTCOffset}{UTC offset in number of hours that the local normal time is 
ahead of UTC, e.g. +01 hour for Berlin nomral (= winter) time}
}
\description{
localNormalTimeToUTC
}
\keyword{internal}
