% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeMain.R
\name{matchingTimeFormat}
\alias{matchingTimeFormat}
\title{Find Time Format matching a Timestamp}
\usage{
matchingTimeFormat(timestamp, timeFormats)
}
\arguments{
\item{timestamp}{character timestamp to be checked against different possible
timstamp formats}

\item{timeFormats}{vector of possible time formats with placeholders %Y
(year), %m (month), %d (day), %H (hours), %M (minutes), %S (seconds) as
described for \code{format.POSIXct}}
}
\value{
first time format in \emph{timeformats} that matches the given 
  \emph{timestamp}. NULL is returned if none of the given \emph{timeformats} 
  matches.
}
\description{
Find Time Format matching a Timestamp
}
\examples{
# Convert a character timestamp of which the format can be one of two
# possible formats into a POSIXct-object
possibleFormats <- c("\%d.\%m.\%Y", "\%d/\%m/\%Y")

x <- "14.01.2015"
t1 <- hsToPosix(x, format = matchingTimeFormat(x, possibleFormats))
  
# In fact this is what stringToPosix does (for only one timestamp)
t2 <- stringToPosix(x, formats = possibleFormats)
stopifnot(identical(t1, t2))
  
# You get a warning if none of the possible formats matches
matchingTimeFormat("01.14.2015", possibleFormats)

}
