% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{minTimeStep}
\alias{minTimeStep}
\title{Minimum Time Step in Sequence of Timestamps}
\usage{
minTimeStep(tstamps, dbg = FALSE)
}
\arguments{
\item{tstamps}{vector of POSIX-type timestamps or any other vector that can
be converted to integer}

\item{dbg}{should debug messages be shown?}
}
\description{
Minimum Time Step in Sequence of Timestamps
}
\examples{
tstamps <- seq(as.POSIXct("2017-11-03"), as.POSIXct("2017-11-04"), 3600)
minTimeStep(tstamps)

# No need for timestamps!
minTimeStep(c(10, 20, 30, 40, 45, 50, 60))
minTimeStep(c(10, 20, 30, 40, 45, 50, 60), dbg = TRUE)

}
