% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{reformatTimestamp}
\alias{reformatTimestamp}
\title{Convert Timstamp String from one Format to another}
\usage{
reformatTimestamp(x, old.format = NULL, new.format = NULL)
}
\arguments{
\item{x}{vector of timestamps (character)}

\item{old.format}{format in which timestamps in \code{x} are given. Default:
\code{"\%Y-\%m-\%d \%H:\%M:\%S"}}

\item{new.format}{format to be applied to timestamps. Default: 
\code{"\%Y-\%m-\%d \%H:\%M:\%S"}}
}
\description{
Convert Timstamp String from one Format to another
}
