% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeMain.R
\name{resetTimePart}
\alias{resetTimePart}
\title{Reduce a Timestamp to a certain Information}
\usage{
resetTimePart(tstamp, resetYear = FALSE, resetMonth = FALSE,
  resetDay = FALSE, resetTime = FALSE, resetHour = resetTime,
  resetMin = resetTime, resetSec = resetTime, ...)
}
\arguments{
\item{tstamp}{(vector of) timestamp(s) of type POSXIXt}

\item{resetYear}{logical. If \code{TRUE}, the year of all timestamps is set
to 1970}

\item{resetMonth}{logical. If \code{TRUE}, the month of all timestamps is set
to 01}

\item{resetDay}{logical. If \code{TRUE}, the day of all timestamps is set to
01}

\item{resetTime}{logical. Used as default for \code{resetHour}, 
\code{resetMin}, \code{resetSec}}

\item{resetHour}{logical. If \code{TRUE}, the hours of all timestamps is set
to 00}

\item{resetMin}{logical. If \code{TRUE}, the minutes of all timestamps is set
to 00}

\item{resetSec}{logical. If \code{TRUE}, the seconds of all timestamps is set
to 00}

\item{\dots}{further arguments to be passed to hsToPosix}
}
\description{
Reduce a Timestamp to a certain Information
}
\examples{
resetTimePart(as.POSIXct("2017-11-04 16:18:44"), resetTime = TRUE)

# Reset year, month and day
resetTimePart(as.POSIXct("2017-11-04 16:18:44"), TRUE, TRUE, TRUE)

}
\keyword{internal}
