% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeStep.R
\name{sequenceOfTimestamps}
\alias{sequenceOfTimestamps}
\title{Create a Sequence of Timestamps}
\usage{
sequenceOfTimestamps(from, to, step.s = 60)
}
\arguments{
\item{from}{first timestamp of mode character in ISO-Syntax: yyyy-mm-dd
[HH:MM:SS] where the part in brackets is optional.}

\item{to}{last timestamp of  mode character in ISO-Syntax: yyyy-mm-dd
[HH:MM:SS] where the part in brackets is optional.}

\item{step.s}{time step between the timestamps in seconds.}
}
\value{
Vector of character timestamps
}
\description{
Creates timestamps of mode character between first timestamp \emph{from} and
\emph{to} with a distance of \emph{step.s} seconds between the timestamps.
}
\examples{
# Create timestamps of January 2011 with five minutes step
sequenceOfTimestamps("2011-01-01 19:00:00", "2011-01-02", 300)

}
