% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{stringToPosix}
\alias{stringToPosix}
\title{Convert a Time String to a POSIXct Object}
\usage{
stringToPosix(x, formats = c("\%Y-\%m-\%d \%H:\%M:\%S", "\%Y-\%m-\%d \%H:\%M",
  "\%Y-\%m-\%d"), ...)
}
\arguments{
\item{x}{character vector of length one representing a timestamp}

\item{formats}{vector of allowed time formats (using \%-placeholders)}

\item{\dots}{arguments passed to \code{\link{hsToPosix}}}
}
\description{
Convert a time string to a POSIXct object. Allow for different possible 
timestamp formats.
}
\examples{
stringToPosix("2016-05-26")
stringToPosix("2016-05-26 12:00")
  
# additional arguments passed to hsToPosix
stringToPosix("2016-05-26 12:00:33", tzone = "ETC/Gmt-1")
  
# lt = TRUE -> create POSIXlt instead of POSIXct
lt1 <- stringToPosix("2016-05-26 17:00", lt = TRUE)
lt2 <- stringToPosix("2016-05-26 17:00", lt = TRUE, tz = "Europe/Berlin")
  
lt1$hour
lt1$isdst # normal time (is daylight saving time = FALSE)
lt2$isdst # summer time (is daylight saving time = TRUE)

}
