% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datetime.R
\name{test_TimeConversionFunctions}
\alias{test_TimeConversionFunctions}
\title{Test Time Conversion Functions}
\usage{
test_TimeConversionFunctions(year = 2000, normalToSummer = TRUE)
}
\arguments{
\item{year}{year for which to demonstrate the switch between Central European
Time (CET) and Central European Summer Time (CEST)}

\item{normalToSummer}{\code{TRUE}: CET to CEST, \code{FALSE}: CEST to CET}
}
\description{
Test Time Conversion Functions
}
