% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{timeFormatToRegex}
\alias{timeFormatToRegex}
\title{Time Format String to Regular Expression Pattern}
\usage{
timeFormatToRegex(timeformat)
}
\arguments{
\item{timeformat}{time format string as used in R's format function with
placeholders such as "%y" (year), "%m" (month), "%d" (day), "%H" (hour),
etc., see \code{\link{as.POSIXct}}}
}
\description{
Convert a time format string to a regular expression pattern that can be used
to check if a date or time string complies with a certain format.
}
\keyword{internal}
