% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesAroundClockChange.R
\name{timesAroundClockChange}
\alias{timesAroundClockChange}
\title{Times Around Clock Change in Central Europe}
\usage{
timesAroundClockChange(
  year = 2000,
  normalToSummer = TRUE,
  step_s = 1800,
  length.out = 5
)
}
\arguments{
\item{year}{year for which to demonstrate the switch between Central European
Time (CET) and Central European Summer Time (CEST)}

\item{normalToSummer}{\code{TRUE}: CET to CEST, \code{FALSE}: CEST to CET}

\item{step_s}{time step in seconds}

\item{length.out}{number of time objects in returned vector}
}
\value{
vector of POSIXct objects with length \code{length.out}
}
\description{
Sequence of Times Around Clock Change in Central Europe
}
\examples{
timesAroundClockChange(2019, normalToSummer = TRUE)
timesAroundClockChange(2019, normalToSummer = FALSE)
timesAroundClockChange(2019, TRUE, step_s = 1, length.out = 3)
timesAroundClockChange(2019, FALSE, step_s = 1, length.out = 3)
}
