% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{timestampIn}
\alias{timestampIn}
\title{Timestamps within Time Interval?}
\usage{
timestampIn(
  tstamps,
  tsFirst = NULL,
  tsLast = NULL,
  firstIncluded = TRUE,
  lastIncluded = FALSE,
  dbg = FALSE
)
}
\arguments{
\item{tstamps}{vector of timestamps, either as character strings in 
ISO-format (yyyy-mm-dd HH:MM:SS), as Date objects or as POSIXt-objects}

\item{tsFirst}{first timestamp (\code{firstIncluded} == TRUE) or timestamp 
directly before first timestamp (\code{firstIncluded} == FALSE) to be 
considered}

\item{tsLast}{last timestamp (\code{lastIncluded} == TRUE) or timestamp 
directly after last timestamp (\code{lastIncluded} == FALSE) to be 
considered}

\item{firstIncluded}{if \code{TRUE}, \code{tsFirst} represents the first,
otherwise the timestamp direcly before the first timestamp to be
considered.}

\item{lastIncluded}{if \code{TRUE}, \code{tsLast} represents the last,
otherwise the timestamp direcly after the last timestamp to be considered.}

\item{dbg}{should debug messages be shown?}
}
\description{
Returns vector of booleans indicating whether the timestamps at corresponding
positions in a vector of timestamps lie within a time interval, specified by 
first and/or last timestamp of the interval.
}
\examples{
tstamps <- sequenceOfTimestamps("2017-11-03", "2017-11-04")
table(timestampIn(tstamps, "2017-11-03 12:00:00", "2017-11-03 13:00:00"))

}
