% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{to.GMT.plus.1}
\alias{to.GMT.plus.1}
\title{Character Timestamps to POSIXct Objects (GMT+1)}
\usage{
to.GMT.plus.1(timestamp)
}
\arguments{
\item{timestamp}{character timestamp(s) to be converted to POSIXct in
timezone "Etc/GMT+1"}
}
\value{
vector of POSIXct in timezone \dQuote{Etc/GMT+1}
}
\description{
Character Timestamps to POSIXct Objects (GMT+1)
}
