% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{toGmtRelativePosix}
\alias{toGmtRelativePosix}
\title{Character Timestamps to POSIXct Objects (GMT+\emph{offset})}
\usage{
toGmtRelativePosix(timestamp, GMT.offset = 1, format = NULL)
}
\arguments{
\item{timestamp}{vector of timestamps (character)}

\item{GMT.offset}{offset to GMT time. Default: 1 = Berlin Normal Time}

\item{format}{format string describing the format of \emph{timstamp}, see
help for \code{strptime}. Default: "\%Y-\%m-\%d \%H:\%M:\%S"}
}
\description{
Convert character timestamps to POSIXct objects in time zont 
GMT+\emph{offset}
}
