% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{toTimezone}
\alias{toTimezone}
\title{Convert POSIXt Object to given Time Zone}
\usage{
toTimezone(x, tz)
}
\arguments{
\item{x}{object of class POSIXt (either POSIXct or POSIXlt)}

\item{tz}{timezone, e.g. "Europe/Berlin", see \code{base::timezones} for help
about possible time zone strings}
}
\value{
POSIXt object in requested timezone (hopefully!)
}
\description{
Convert POSIXt Object to given Time Zone
}
\keyword{internal}
