% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{toUTC}
\alias{toUTC}
\title{Convert POSIXt Object to UTC Time Zone}
\usage{
toUTC(x)
}
\arguments{
\item{x}{object of class POSIXt (either POSIXct or POSIXlt)}
}
\value{
POSIXt object in UTC timezone (hopefully!)
}
\description{
Convert POSIXt Object to UTC Time Zone
}
\examples{

# Create a timestamp in the time zone that is set on the local machine
time <- as.POSIXct("2017-11-01 01:16")

# Convert time zone to UTC
time_utc <- toUTC(time)

# The new time zone "UTC" is set in the attribute "tz"
attr(time_utc, "tz")

# The times mean the same, just expressed in another time zone!
time_utc == time

}
