% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc_offset_berlin.R
\name{utc_offset_Berlin_time}
\alias{utc_offset_Berlin_time}
\title{UTC Offsets of Berlin Local Timestamps}
\usage{
utc_offset_Berlin_time(timestamps)
}
\arguments{
\item{timestamps}{vector of character representing timestamps in format
\code{"yyyy-mm-dd HH:MM:SS"}}
}
\value{
vector of elements \code{"+0100"} or \code{"+0200"}, depending on 
  whether the timestamps at corresponding positions in \code{timestamps} are
  in CET or CEST, respectively.
}
\description{
For local timestamps (character) in the format \code{"yyyy-mm-dd HH:MM:SS"},
of which is known that they are recorded in time zone Europe/Berlin, i.e. CET
in winter and CEST in summer, the UTC offset (i.e. \code{"+0100"} in winter
and \code{"+0200"} in summer) is determined. Therfore, it is required that
the \code{timestamps} are ordered by time, which should be the case if they 
were recorded by a measuring device. Use this function to create unique 
timestamps by adding their UTC offset.
}
\examples{
# Change from CET to CEST
utc_offset_Berlin_time(c(
  "2017-03-26 01:58:00",
  "2017-03-26 01:59:00",
  "2017-03-26 03:00:00", # jump from 02:00 to 03:00
  "2017-03-26 03:01:00",
  "2017-03-26 03:02:00"
))
  
#> "+0200" "+0200" "+0100" "+0100" "+0100"

# Note that the following timestamps do in fact not exist but do
# nevertheless result in "+0200"
utc_offset_Berlin_time(c(
  "2017-03-26 02:00:00",
  "2017-03-26 02:15:00",
  "2017-03-26 02:30:00",
  "2017-03-26 02:45:00"
))

#> "+0200" "+0200" "+0200" "+0200"

# Change from CEST to CET
utc_offset_Berlin_time(c(
  "2017-10-29 01:30:00", # CEST
  "2017-10-29 02:00:00", # first time: CEST
  "2017-10-29 02:30:00", # first time: CEST
  "2017-10-29 02:00:00", # second time: CET
  "2017-10-29 02:30:00"  # second time: CET
))
  
 #> "+0200" "+0200" "+0200" "+0100" "+0100"

}
