% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc_offset_berlin.R
\name{utc_offset_Berlin_time_1d}
\alias{utc_offset_Berlin_time_1d}
\title{UTC Offsets of Berlin Local Timestamps within one Day}
\usage{
utc_offset_Berlin_time_1d(x)
}
\arguments{
\item{x}{vector of character representing timestamps in format 
\code{"yyyy-mm-dd HH:MM:SS"}. All timestamps must belong to one and the
same day.}
}
\value{
vector of elements \code{"+0100"} or \code{"+0200"}, depending on 
  whether the timestamps at corresponding positions in \code{x} are
  in CET or CEST, respectively.
}
\description{
UTC Offsets of Berlin Local Timestamps within one Day
}
\examples{

# At what days does Central European Summer Time (CEST) start/end?
cest_begin_end <- kwb.datetime::date_range_CEST(2019)

# Provide text timestamps "around" the switch from CET to CEST
times_cet_cest <- c(
  "2019-03-31 01:00:00", # CET
  "2019-03-31 01:15:00", # CET
  "2019-03-31 01:30:00", # CET
  "2019-03-31 01:45:00", # CET
  "2019-03-31 03:00:00", # CEST
  "2019-03-31 03:15:00", # CEST
  "2019-03-31 03:30:00"  # CEST
)

# Provide text timestamps "around" the switch from CEST to CET
times_cest_cet <- c(
  "2019-10-27 01:00:00", # CEST
  "2019-10-27 01:30:00", # CEST
  "2019-10-27 02:00:00", # CEST
  "2019-10-27 02:30:00", # CEST
  "2019-10-27 02:00:00", # CET
  "2019-10-27 02:30:00", # CET
  "2019-10-27 03:00:00", # CET
  "2019-10-27 03:30:00", # CET
  "2019-10-27 04:00:00"  # CET
)

# Get the offset strings
offsets_cet_cest <- kwb.datetime:::utc_offset_Berlin_time_1d(times_cet_cest)
offsets_cest_cet <- kwb.datetime:::utc_offset_Berlin_time_1d(times_cest_cet)

# Create ISO norm timestamps including the offset
iso_cet_cest <- paste0(times_cet_cest, offsets_cet_cest)
iso_cest_cet <- paste0(times_cest_cet, offsets_cest_cet)

# Use the function iso_to_localtime() to create POSIXct-objects in Europe/Berlin
kwb.datetime:::iso_to_localtime(iso_cet_cest)
kwb.datetime:::iso_to_localtime(iso_cest_cet)

}
\seealso{
\code{\link{utc_offset_Berlin_time}}
}
\keyword{internal}
