% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{utcOffset}
\alias{utcOffset}
\title{Get UTC Offset from Local and UTC Timestamp (Character)}
\usage{
utcOffset(LocalDateTime, DateTimeUTC)
}
\arguments{
\item{LocalDateTime}{character string representing a local timestamp}

\item{DateTimeUTC}{character string representing a time stamp in UTC}
}
\description{
Get UTC Offset from Local and UTC Timestamp (Character)
}
\keyword{internal}
