% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{utcToBerlinLocalTime}
\alias{utcToBerlinLocalTime}
\title{utcToBerlinLocalTime}
\usage{
utcToBerlinLocalTime(x)
}
\arguments{
\item{x}{Character timestamp that is interpretable by
\code{\link{as.POSIXct}}, representing a time given in UTC}
}
\description{
utcToBerlinLocalTime
}
\keyword{internal}
