% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEqualStepRanges.R
\name{getEqualStepRanges}
\alias{getEqualStepRanges}
\title{Sequences of Date Time Objects With Equal Time Step}
\usage{
getEqualStepRanges(times)
}
\arguments{
\item{times}{vector of \code{POSIXct} objects}
}
\description{
Sequences of Date Time Objects With Equal Time Step
}
\examples{
# Generate a sequence of date and time objects
as_berlin_posix <- function(x) as.POSIXct(x, tz = "Europe/Berlin")
times <- seq(
  from = as_berlin_posix("2019-01-01"), 
  to = as_berlin_posix("2020-01-01"), 
  by = 3600
)

# As expected, exactly one sequence of equal time step is found:
getEqualStepRanges(times)

# Simulate the case that timestamps were read from a text file and converted
# with as.POSIXct()
timestamps <- as.character(times)
new_times <- as.POSIXct(timestamps, tz = "Europe/Berlin")

# Show the sequences of equal time steps again
getEqualStepRanges(new_times)

# What happened? The timestamp 2019-10-27 02:00 appears twice! Once in CEST
# and once in CET. Use a helper function that assigns CEST and CET as 
# required:
good_times <- textToEuropeBerlinPosix(timestamps)

# Check if the original date and time objects could be reproduced
identical(good_times, times)

}
