% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchingTimeFormat.R
\name{hasTimeFormat}
\alias{hasTimeFormat}
\title{Do Timestamps have the expected Format?}
\usage{
hasTimeFormat(timestamps, timeformat = NULL)
}
\arguments{
\item{timestamps}{character strings representing timestamps.}

\item{timeformat}{character string giving a date-time format as used by 
\code{\link{strptime}}.}
}
\description{
Checks if timestamps are formatted according to timeformat
}
\examples{
hasTimeFormat("1.1.2012", "\%d.\%m.\%Y") # TRUE
hasTimeFormat("1.13.2012", "\%d.\%m.\%Y") # FALSE
hasTimeFormat("1/31/2012", "\%m/\%d/\%Y") # TRUE
hasTimeFormat("31/1/2012", "\%m/\%d/\%Y") # FALSE

hasTimeFormat(c("1.1.", "1.13.", "12.12.", "32.1."), "\%d.\%m.") 
#               TRUE     FALSE    TRUE      FALSE

}
