% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeConvert.R
\name{toUTC}
\alias{toUTC}
\title{Convert POSIXt Object to UTC Time Zone}
\usage{
toUTC(x)
}
\arguments{
\item{x}{object of class POSIXt (either POSIXct or POSIXlt)}
}
\value{
POSIXt object in UTC timezone (hopefully!)
}
\description{
Convert POSIXt Object to UTC Time Zone
}
