% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc_offset_berlin.R
\name{utc_offset_Berlin_day}
\alias{utc_offset_Berlin_day}
\title{UTC Offsets of Days in Time Zone Europe/Berlin}
\usage{
utc_offset_Berlin_day(x)
}
\arguments{
\item{x}{vector of character representing date strings in the format
\code{yyyy-mm-dd}}
}
\value{
vector of elements \code{"+01"} or \code{"+02"}, depending on 
  whether all timestamps of the days at corresponding positions in \code{x}
  are in winter (CET) or summer (CEST), respectively. For days at which the
  time is adjusted from CET to CEST or vice versa, \code{NA} is returned.
}
\description{
UTC Offsets of Days in Time Zone Europe/Berlin
}
\examples{
kwb.datetime:::utc_offset_Berlin_day("2017-11-04") #> "+01" -> CET

# The offset is not unique at the days of clock change CEST -> CET
kwb.datetime:::utc_offset_Berlin_day("2017-10-28") #> "+02" -> CEST
kwb.datetime:::utc_offset_Berlin_day("2017-10-29") #> NA -> offset not unique!
kwb.datetime:::utc_offset_Berlin_day("2017-10-30") #> "+01" -> CET

# The offset is not unique at the days of clock change CET -> CEST
kwb.datetime:::utc_offset_Berlin_day("2017-03-25") #> "+01" -> CET
kwb.datetime:::utc_offset_Berlin_day("2017-03-26") #> NA -> offset not unique!
kwb.datetime:::utc_offset_Berlin_day("2017-03-27") #> "+02" -> CEST

}
\seealso{
\code{\link{utc_offset_Berlin_time}}
}
\keyword{internal}
