% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc_offset_berlin.R
\name{utc_offset_Berlin_time_1d}
\alias{utc_offset_Berlin_time_1d}
\title{UTC Offsets of Berlin Local Timestamps within one Day}
\usage{
utc_offset_Berlin_time_1d(x)
}
\arguments{
\item{x}{vector of character representing timestamps in format 
\code{"yyyy-mm-dd HH:MM:SS"}. All timestamps must belong to one and the
same day.}
}
\value{
vector of elements \code{"+0100"} or \code{"+0200"}, depending on 
  whether the timestamps at corresponding positions in \code{x} are
  in CET or CEST, respectively.
}
\description{
UTC Offsets of Berlin Local Timestamps within one Day
}
\examples{
kwb.datetime:::utc_offset_Berlin_time_1d(c(
  "2017-10-29 00:00:00",
  "2017-10-29 01:00:00",
  "2017-10-29 02:00:00",
  "2017-10-29 02:00:00",
  "2017-10-29 03:00:00"
))

# The following results in an error because the timestamps 
# belong to diffeernt days
# kwb.datetime:::utc_offset_Berlin_time_1d(c(
#   "2017-10-29 00:00:00", 
#   "2017-10-30 00:00:00"
# ))

}
\seealso{
\code{\link{utc_offset_Berlin_time}}
}
\keyword{internal}
